use components.lifecycle to handle server init using manager
