+++ shortfin/python/shortfin_apps/llm/routes/generate.py
@@ -10,7 +10,6 @@ from shortfin.interop.fastapi import FastAPIResponder
 
 from ..components.generate import ClientGenerateBatchProcess
 from ..components.io_struct import GenerateReqInput
-from ..lifecycle_hooks import services
 
 generation_router = APIRouter()
 
@@ -18,7 +17,7 @@ generation_router = APIRouter()
 @generation_router.post("/generate")
 @generation_router.put("/generate")
 async def generate_request(gen_req: GenerateReqInput, request: Request):
-    service = services["default"]
+    service = request.app.state.services["default"]
     gen_req.post_init()
     responder = FastAPIResponder(request)
     ClientGenerateBatchProcess(service, gen_req, responder).launch()
